+++ tests/unit/common/test_generate_table.py
@@ -228,6 +228,13 @@ class TestDynamicDynamicTable(TestCase):
             self.TestDTRTable(name='test_dtr_table', description='my table',
                               target_tables={'optional_col3': test_table})
 
+    def test_attribute(self):
+        test_table = self.TestTable(name='test_table', description='my test table')
+        assert test_table.my_col is not None
+        assert test_table.indexed_col is not None
+        assert test_table.my_col is test_table['my_col']
+        assert test_table.indexed_col is test_table['indexed_col'].target
+
     def test_roundtrip(self):
         # NOTE this does not use H5RoundTripMixin because this requires custom validation
         test_table = self.TestTable(name='test_table', description='my test table')